+++ core/src/main/java/com/refinedmods/refinedstorage2/core/network/Network.java
@@ -12,6 +12,10 @@ public interface Network {
 
     Set<NetworkNodeReference> getNodeReferences();
 
+    void onNodesChanged();
+
+    EnergyStorage getEnergyStorage();
+
     void invalidateStorageChannelSources();
 
     StorageChannel<Rs2ItemStack> getItemStorageChannel();