+++ core/src/main/java/com/refinedmods/refinedstorage2/core/network/node/controller/ControllerEnergyState.java
@@ -1,0 +1,8 @@
+package com.refinedmods.refinedstorage2.core.network.node.controller;
+
+public enum ControllerEnergyState {
+    OFF,
+    NEARLY_ON,
+    ON,
+    NEARLY_OFF
+}