+++ fabric/src/main/java/com/refinedmods/refinedstorage2/fabric/block/ControllerBlock.java
@@ -33,6 +33,13 @@ public class ControllerBlock extends NetworkNodeBlock {
             return result;
         }
 
+        if (!world.isClient()) {
+            BlockEntity blockEntity = world.getBlockEntity(pos);
+            if (blockEntity instanceof ControllerBlockEntity) {
+                ((ControllerBlockEntity) blockEntity).receive();
+            }
+        }
+
         return super.onUse(state, world, pos, player, hand, hit);
     }