+++ fabric/src/main/java/com/refinedmods/refinedstorage2/fabric/block/entity/ControllerBlockEntity.java
@@ -1,6 +1,8 @@
 package com.refinedmods.refinedstorage2.fabric.block.entity;
 
-import com.refinedmods.refinedstorage2.core.network.node.NetworkNodeImpl;
+import com.refinedmods.refinedstorage2.core.network.node.controller.ControllerNetworkNode;
+import com.refinedmods.refinedstorage2.core.util.Action;
+import com.refinedmods.refinedstorage2.fabric.Rs2Config;
 import com.refinedmods.refinedstorage2.fabric.Rs2Mod;
 import com.refinedmods.refinedstorage2.fabric.block.ControllerBlock;
 import com.refinedmods.refinedstorage2.fabric.block.ControllerEnergyType;
@@ -11,44 +13,40 @@ import com.refinedmods.refinedstorage2.fabric.util.Positions;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
-public class ControllerBlockEntity extends NetworkNodeBlockEntity<NetworkNodeImpl> {
+public class ControllerBlockEntity extends NetworkNodeBlockEntity<ControllerNetworkNode> {
+    private long lastTypeChanged;
+    private ControllerEnergyType lastType = ControllerEnergyType.OFF;
+
     public ControllerBlockEntity() {
         super(Rs2Mod.BLOCK_ENTITIES.getController());
     }
 
-    private int ticks = 0;
-    private ControllerEnergyType type = ControllerEnergyType.OFF;
-
     @Override
     public void tick() {
-        super.tick();
+        if (world != null && !world.isClient() && node != null) {
+            ControllerEnergyType type = ControllerEnergyType.ofState(node.getState());
+            if (type != lastType && (lastTypeChanged == 0 || System.currentTimeMillis() - lastTypeChanged > 1000)) {
+                this.lastTypeChanged = System.currentTimeMillis();
+                this.lastType = type;
 
-        if (world != null && !world.isClient()) {
-            ++ticks;
-
-            if (ticks % 20 == 0) {
-                type = next(type);
                 world.setBlockState(pos, world.getBlockState(pos).with(ControllerBlock.ENERGY_TYPE, type));
             }
         }
     }
 
-    private ControllerEnergyType next(ControllerEnergyType current) {
-        switch (current) {
-            case OFF:
-                return ControllerEnergyType.NEARLY_ON;
-            case NEARLY_OFF:
-                return ControllerEnergyType.OFF;
-            case NEARLY_ON:
-                return ControllerEnergyType.ON;
-            case ON:
-                return ControllerEnergyType.NEARLY_OFF;
-        }
-        return ControllerEnergyType.OFF;
+    @Override
+    protected ControllerNetworkNode createNode(World world, BlockPos pos) {
+        return new ControllerNetworkNode(
+                FabricRs2WorldAdapter.of(world),
+                Positions.ofBlockPos(pos),
+                FabricNetworkNodeReference.of(world, pos),
+                Rs2Config.get().getController().getCapacity()
+        );
     }
 
-    @Override
-    protected NetworkNodeImpl createNode(World world, BlockPos pos) {
-        return new NetworkNodeImpl(FabricRs2WorldAdapter.of(world), Positions.ofBlockPos(pos), FabricNetworkNodeReference.of(world, pos));
+    public void receive() {
+        if (node != null) {
+            node.receive(5, Action.EXECUTE);
+        }
     }
 }