+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Currently, this version only works on Fabric.
 
 ### Added
 
+- Controller
 - Grid
 - Disk Drive
 - Storage Part
@@ -46,3 +47,5 @@ Currently, this version only works on Fabric.
 - The Grid search bar now has much more powerful searching, supporting expressions. Consult the Patchouli documentation.
 - A single item in storage can now be larger than 2,147,483,647.
 - You can now place directional blocks facing up/down.
+- A controller that is in nearly on/nearly off state is no longer fullbright.
+- You can now add multiple controllers to a network to meet the energy requirements of your network.